Fix ClassCastException: Double cannot be cast to BigDecimal on cached exchange rate lookup

+++ src/main/kotlin/com/fintech/currencyconverter/adapter/output/gateway/CurrencyApiGateway.kt
@@ -29,7 +29,7 @@ class CurrencyApiGateway(
             ?: throw ExchangeRateUnavailableException(
                 "Rate not found for $sourceCurrency -> $targetCurrency",
             )
-        return BigDecimal.valueOf(rateData.value)
+        return rateData.value
     }
 
     @Suppress("unused")
